feat: Update testing and session management for Clerk integration
- Modified package.json to streamline Jest configurations for unit and integration tests.
- Removed outdated test files for Clerk's controller and module, consolidating tests for improved clarity.
- Enhanced ElasticsearchService with a new deleteUser method for user management.
- Refactored UsersService to utilize the new deleteUser method for cleaner code.
- Updated E2E tests to reflect the removal of AdminGuard and ensure proper access control with RolesGuard.
- Adjusted jest-e2e.json for better test organization and module resolution.
